Changes requested by Anthony as part of Pull Request.
This is a combination of 11 commits.

Update src/AudioAnalysisTools/Events/SpectralEvent.cs


Update src/AudioAnalysisTools/Events/SpectralEvent.cs


Update src/TowseyLibrary/MatrixTools.cs


Update src/TowseyLibrary/MatrixTools.cs


Update src/AnalysisPrograms/Recognizers/Birds/NinoxBoobook.cs


Update src/AudioAnalysisTools/Events/SpectralEvent.cs


Update src/AudioAnalysisTools/Events/SpectralEvent.cs


Update src/AudioAnalysisTools/Events/SpectralEvent.cs


Update src/AudioAnalysisTools/Events/SpectralEvent.cs


Changes requested by Anthony as part of Pull Request.

Issue #319

Update src/AnalysisPrograms/Recognizers/Birds/NinoxBoobook.cs
